finalizing testing and exception handling

+++ src/main/java/com/revature/servlets/DeleteUserServlet.java
@@ -2,6 +2,8 @@ package com.revature.servlets;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.revature.dtos.ErrorResponse;
+import com.revature.dtos.Principal;
+import com.revature.exceptions.InvalidInputException;
 import com.revature.exceptions.InvalidRequestException;
 import com.revature.services.UserService;
 
@@ -25,10 +27,34 @@ public class DeleteUserServlet extends HttpServlet {
         PrintWriter respWriter = resp.getWriter();
         try {
 
+            String principalJSON = (String) req.getSession().getAttribute("principal");
+
+            if (principalJSON == null) {
+                ErrorResponse err = new ErrorResponse(401, "No principal object found on request.");
+                respWriter.write(mapper.writeValueAsString(err));
+                resp.setStatus(401);
+                return; //make sure it doesnt fall through
+            }
+
+            Principal principal = mapper.readValue(principalJSON, Principal.class);
+
+            if (!principal.getRole().equalsIgnoreCase("Admin")) {
+                ErrorResponse err = new ErrorResponse(403, "Forbidden: your role does not permit you to access this endpoint.");
+                respWriter.write(mapper.writeValueAsString(err));
+                resp.setStatus(403);
+                return;
+            }
+
             String deletedUserId = req.getParameter("id");
             System.out.println(deletedUserId);
             if (deletedUserId != "null") {
                 int id = Integer.parseInt(deletedUserId);
+                if (id == principal.getId()) {
+                    ErrorResponse err = new ErrorResponse(403, "Forbidden: You cannot delete yourself...");
+                    respWriter.write(mapper.writeValueAsString(err));
+                    resp.setStatus(403);
+                    return;
+                }
                 userService.deleteUser(id);
                 resp.setStatus(204);
             } else {
@@ -38,8 +64,6 @@ public class DeleteUserServlet extends HttpServlet {
             }
 
 
-
-
         } catch (NumberFormatException | InvalidRequestException mie) {
             mie.printStackTrace();
             resp.setStatus(400);
@@ -47,7 +71,13 @@ public class DeleteUserServlet extends HttpServlet {
             ErrorResponse err = new ErrorResponse(400, "Bad Req: Malform int object found in request body");
             String errJSON = mapper.writeValueAsString(err);
             respWriter.write(errJSON);
+        } catch (InvalidInputException iie) {
+            iie.printStackTrace();
+            resp.setStatus(409);
 
+            ErrorResponse err = new ErrorResponse(400, "This user does not exist, therefore cannot be deleted.");
+            String errJSON = mapper.writeValueAsString(err);
+            respWriter.write(errJSON);
         } catch (Exception e) {
             e.printStackTrace();
             resp.setStatus(500); // 500 = INTERNAL SERVER ERROR
